+++ CHANGELOG.md
@@ -1,1 +1,29 @@
+Changelog
+=========
 
+This log should always be updated when doing backwards incompatible changes, resulting in a major version bump. Feel free to add a log for lesser version bumps as well, but for major bumps it's a must.
+
+## 2.0.0
+
+Major revision of the project, switched from Dash/Flask to Django.
+
+### Added:
+
+Numerous new explorer features:
+
+* content tables
+* multiindex columns in frequencies table
+* n-grams
+* frequency table for whole corpus
+* etc.
+
+Other site features:
+
+* Login
+* User preferences
+* etc.
+
+
+## 1.x.x
+
+Dash/Flask version of the app; prototype only. Using this is not advised.